Refine c1z save path portability
